tests/wpt/web-platform-tests/css/css-images/parsing/image-resolution-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Images Module Level 4: parsing image-resolution with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-images-4/#propdef-image-resolution">
<meta name="assert" content="image-resolution supports the full grammar '[ from-image || <resolution> ] && snap?'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>



test_valid_value("image-resolution", "1dpi");
test_valid_value("image-resolution", "2dpcm from-image");
test_valid_value("image-resolution", "3dppx from-image snap");
test_valid_value("image-resolution", "4dpi snap");
test_valid_value("image-resolution", "from-image");
test_valid_value("image-resolution", "from-image 5dpcm");
test_valid_value("image-resolution", "from-image 6dppx snap");
test_valid_value("image-resolution", "from-image snap");
test_valid_value("image-resolution", "snap 7.5dpi");
test_valid_value("image-resolution", "snap -8dpcm from-image");
test_valid_value("image-resolution", "snap from-image");
test_valid_value("image-resolution", "snap from-image 0dppx");


</script>
</body>
</html>
