third_party/blink/web_tests/external/wpt/css/css-images/parsing/image-resolution-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['image-resolution'] = "1dpi" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['image-resolution'] = "2dpcm from-image" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['image-resolution'] = "3dppx from-image snap" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['image-resolution'] = "4dpi snap" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['image-resolution'] = "from-image" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['image-resolution'] = "from-image 5dpcm" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['image-resolution'] = "from-image 6dppx snap" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['image-resolution'] = "from-image snap" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['image-resolution'] = "snap 7.5dpi" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['image-resolution'] = "snap -8dpcm from-image" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['image-resolution'] = "snap from-image" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['image-resolution'] = "snap from-image 0dppx" should set the property value
  assert_not_equals: property should be set got disallowed value ""
Harness: the test ran to completion.
